Add base shooting.(Yhuuu)

+++ Story_One_Coube/Models/Point.cs
@@ -8,10 +8,10 @@ namespace Story_One_Coube.Models
 {
     class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             X = x;
             Y = y;
